Create SegmentGroup, remove logs, move glossary docs to AnnotatedText
